+++ doc/examples/list_files_in_bucket.py
@@ -1,8 +1,8 @@
 from pathlib import Path
-from exasol_bucketfs_util_python import upload, list_files
-from exasol_bucketfs_util_python.bucket_config import BucketConfig
-from exasol_bucketfs_util_python.bucketfs_config import BucketFSConfig
-from exasol_bucketfs_util_python.bucketfs_connection_config import BucketFSConnectionConfig
+from exasol_bucketfs_utils_python import upload, list_files
+from exasol_bucketfs_utils_python.bucket_config import BucketConfig
+from exasol_bucketfs_utils_python.bucketfs_config import BucketFSConfig
+from exasol_bucketfs_utils_python.bucketfs_connection_config import BucketFSConnectionConfig
 
 connection_config = BucketFSConnectionConfig(
     host="localhost", port=6666,
