+++ doc/examples/upload_github_release_to_bucket.py
@@ -1,4 +1,4 @@
-from exasol_bucketfs_util_python.github_release_file_bucketfs_uploader import GithubReleaseFileBucketFSUploader
+from exasol_bucketfs_utils_python.github_release_file_bucketfs_uploader import GithubReleaseFileBucketFSUploader
 
 release_uploader = \
     GithubReleaseFileBucketFSUploader(file_to_download_name="file",
