+++ exasol_bucketfs_utils_python/abstract_bucketfs_location.py
