+++ exasol_bucketfs_utils_python/bucketfs_connection_config.py
