+++ exasol_bucketfs_utils_python/github_release_file_bucketfs_uploader.py
@@ -1,7 +1,7 @@
 import requests
 from requests.auth import HTTPBasicAuth
 
-from exasol_bucketfs_util_python.release_link_extractor import ReleaseLinkExtractor
+from exasol_bucketfs_utils_python.release_link_extractor import ReleaseLinkExtractor
 
 
 class GithubReleaseFileBucketFSUploader:
