+++ exasol_bucketfs_utils_python/list_files.py
@@ -1,9 +1,9 @@
 from typing import Iterable
 import requests
 from pathlib import Path
-from exasol_bucketfs_util_python.bucket_config import BucketConfig
-from exasol_bucketfs_util_python import bucketfs_utils
-from exasol_bucketfs_util_python.bucketfs_utils import generate_bucket_http_url
+from exasol_bucketfs_utils_python.bucket_config import BucketConfig
+from exasol_bucketfs_utils_python import bucketfs_utils
+from exasol_bucketfs_utils_python.bucketfs_utils import generate_bucket_http_url
 
 
 def list_files_in_bucketfs(bucket_config: BucketConfig,
