+++ exasol_bucketfs_utils_python/upload.py
@@ -4,9 +4,9 @@ from typing import Tuple, IO, Any
 from urllib.parse import ParseResult
 import joblib
 import requests
-from exasol_bucketfs_util_python import bucketfs_utils
-from exasol_bucketfs_util_python.bucket_config import BucketConfig
-from exasol_bucketfs_util_python.bucketfs_utils import \
+from exasol_bucketfs_utils_python import bucketfs_utils
+from exasol_bucketfs_utils_python.bucket_config import BucketConfig
+from exasol_bucketfs_utils_python.bucketfs_utils import \
     generate_bucket_http_url, generate_bucket_udf_path
