+++ tests/fixtures/prepare_bucket_fixture.py
@@ -1,8 +1,8 @@
 import pytest
-from exasol_bucketfs_util_python import upload
-from exasol_bucketfs_util_python.bucket_config import BucketConfig
-from exasol_bucketfs_util_python.bucketfs_config import BucketFSConfig
-from exasol_bucketfs_util_python.bucketfs_connection_config import \
+from exasol_bucketfs_utils_python import upload
+from exasol_bucketfs_utils_python.bucket_config import BucketConfig
+from exasol_bucketfs_utils_python.bucketfs_config import BucketFSConfig
+from exasol_bucketfs_utils_python.bucketfs_connection_config import \
     BucketFSConnectionConfig
 from tests.integration_tests.with_db.test_load_fs_file_from_udf import delete_testfile_from_bucketfs
