+++ tests/integration_tests/without_db/test_delete_file.py
@@ -1,9 +1,9 @@
 import pytest
 
-from exasol_bucketfs_util_python import upload, list_files, delete
-from exasol_bucketfs_util_python.bucket_config import BucketConfig
-from exasol_bucketfs_util_python.bucketfs_config import BucketFSConfig
-from exasol_bucketfs_util_python.bucketfs_connection_config import BucketFSConnectionConfig
+from exasol_bucketfs_utils_python import upload, list_files, delete
+from exasol_bucketfs_utils_python.bucket_config import BucketConfig
+from exasol_bucketfs_utils_python.bucketfs_config import BucketFSConfig
+from exasol_bucketfs_utils_python.bucketfs_connection_config import BucketFSConnectionConfig
 from tests.integration_tests.with_db.test_load_fs_file_from_udf import delete_testfile_from_bucketfs
