+++ tests/integration_tests/without_db/test_upload_download.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 
-from exasol_bucketfs_util_python import upload, download
-from exasol_bucketfs_util_python.bucket_config import BucketConfig
-from exasol_bucketfs_util_python.bucketfs_config import BucketFSConfig
-from exasol_bucketfs_util_python.bucketfs_connection_config import BucketFSConnectionConfig
+from exasol_bucketfs_utils_python import upload, download
+from exasol_bucketfs_utils_python.bucket_config import BucketConfig
+from exasol_bucketfs_utils_python.bucketfs_config import BucketFSConfig
+from exasol_bucketfs_utils_python.bucketfs_connection_config import BucketFSConnectionConfig
 
 
 def test_file_upload_download():
