+++ tests/unit_tests/test_bucketfs_utils.py
@@ -1,9 +1,9 @@
 import pytest
 
-from exasol_bucketfs_util_python import bucketfs_utils
-from exasol_bucketfs_util_python.bucketfs_config import BucketFSConfig
-from exasol_bucketfs_util_python.bucketfs_connection_config import BucketFSConnectionConfig
-from exasol_bucketfs_util_python.bucket_config import BucketConfig
+from exasol_bucketfs_utils_python import bucketfs_utils
+from exasol_bucketfs_utils_python.bucketfs_config import BucketFSConfig
+from exasol_bucketfs_utils_python.bucketfs_connection_config import BucketFSConnectionConfig
+from exasol_bucketfs_utils_python.bucket_config import BucketConfig
 
 
 def test_generate_bucket_udf_path_non_archive_file():
